Added battery guardian
There seems to be a problem with the charging in my model: after
a day with the battery fully charged, it starts to loose charge,
but switching each 10-20 second between 100%charged, not charging
and X% charged, charging (with X a descending value).

This patch adds a guardian that checks for this situation, and if
it detects it, orders the robot to start cleaning to send it away
of the charging base, and then send it again to charge. This seems
to fix the problem.

It is configurable in the persistent data with the keys:

battery_guard_enabled: a boolean value that enables or disables
   it. True by default

battery_guard_level: the battery level to detect this problem.
   If the battery charge is lower than this value, it will
   be supposed that it is happening. By default it is 80 (it
   is a percentage; thus between 0 and 100).

battery_guard_times: the number of times that the sequence
   100% level, not charging -> <level%, charging has to
   happen before activating the safeguard. By default it is
   3, to avoid false cases where the battery is really
   charging.