Don't need to use dnsmasq
Thanks to Daniel Webber, who tested it, we confirmed that it is
possible to pass the server IP when pairing the robot. This
removes the need of mounting a DNS server and masking the
original DNS entries from the manufacturer.

This patch implements all this: now the pairconga.py and the
configconga.py programs also ask for the server IP, and it's
no longer mandatory to install or configure DNSMASQ.